+++ src/main/java/project/persistence/repositories/UserExerciseRepository.java
@@ -1,7 +1,6 @@
 package project.persistence.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import project.persistence.entities.UserExercise;
 
 import java.util.List;
@@ -14,6 +13,6 @@ public interface UserExerciseRepository extends JpaRepository<UserExercise, Long
 
     UserExercise findOne(Long id);
 
-    List<UserExercise> findByUserID(int userID);
+    List<UserExercise> findByUserID(Long userID);
 
 }
