+++ src/main/java/project/service/Implementation/UserExerciseServiceImplementation.java
@@ -53,7 +53,7 @@ public class UserExerciseServiceImplementation implements UserExerciseService {
     }
 
     @Override
-    public List<UserExercise> findByUserID(int userID) {
+    public List<UserExercise> findByUserID(Long userID) {
         return repository.findByUserID(userID);
     }
